+++ src/lib.rs
@@ -88,8 +88,8 @@ use oxideav_core::{CodecCapabilities, CodecId, MediaType};
 use oxideav_core::{CodecInfo, CodecRegistry};
 
 pub use ass_anim::{
-    animate_style_at, collect_transforms, fade_alpha_at, karaoke_fills, position_at,
-    transform_factor, FadeAlpha, KaraokeSyllable,
+    animate_style_at, collect_transforms, evaluate_line_at, fade_alpha_at, karaoke_fills,
+    position_at, transform_factor, EvaluatedLine, EvaluatedSpan, FadeAlpha, KaraokeSyllable,
 };
 pub use ass_emit::{color_to_string, serialize_line, style_row_to_string};
 pub use ass_event::{